consolidate alternatives
the current implemention uses the title to create the alternative
wrapper. for the result of the method the title is "methodObjectResult",
but the alternatives are ContentDescriptor and Reference which are the
same as ContentDescriptorOrReference from e.g. params. The new structure
allow to define ONE instance and use them as param and result.